+++ lib/mongoid-store.rb
@@ -1,3 +1,3 @@
-require 'active_support/cache/mongoid-store'
+require 'active_support/cache/mongoid_store'
 
 module MongoidStore; end
